azure-functions/requirements.txt
# Azure Functions Python Requirements - Minimal for testing

# Azure Functions
azure-functions